Fix bugs uncovered by compiler warnings
clang 3.5 and GCC 4.9 have been used with -Wall -Werror


git-svn-id: https://bioproj.extra.cea.fr/neurosvn/brainvisa/aims/aims-free/branches/bug_fix@69235 613412ce-1361-4b7a-93bf-a3c75af1bae4

+++ aimsalgo/src/aimsalgo/scalespace/meshDiffuse.cc
@@ -186,7 +186,7 @@ void TextureSmoothing::surfaceBlockSmooth( const string & latt_filename,
               /* WARNING only default type "int" */
 
               fseek( fp_neighs, 4 * k * nodes_block * ncolPINV, 0 );
-              int len = fread( IND_NEIGHS_chunk, sizeof(int),
+              size_t len = fread( IND_NEIGHS_chunk, sizeof(int),
                      nodes_block_curr * ncolPINV, fp_neighs );
               if( len != nodes_block_curr*ncolPINV )
               {
